requirements/dev.txt
mysqlclient
requests
redis
lxml
prospector
nose
pyyaml>=5.1
django
celery>=4.2
djangorestframework==3.7.3
django-filter
django-extensions
djangorestframework-jwt
django-rest-swagger
django-cas-client==1.5.0
django-cors-headers
django-celery-beat
django-redis
pinyin
clocwalk>=0.2.0
django-debug-toolbar
django-conn-pool
markdownify
gevent
rsa
func_timeout
markdown